PG18: fix query results diff in merge regress test.
The test used SQL functions which can be cached in PG18+ since
commit 0dca5d68d. This means that a Citus distributed plan can
be cached, but the copy function for a distributed plan did not
include the `sourceResultRepartitionColumnIndex` field, which is
critical for MERGE queries. Including it fixes the diff problem.
Also update `OutDistributedPlan()` function so that the field is
picked up by `nodeToString()`.